#73 Fix audience view not loading, also got sidetracked and changed the grid action drawer to have buttons so they can be tabbed through
