+++ src/client/src/api/audiences/audienceResponse.ts
@@ -2,6 +2,7 @@ export interface AudienceResponse {
   id: string
   name: string
   url: string
+  isRelated: boolean
   isEditable: boolean
   isDeletable: boolean
   isLinkableToSchedule: boolean
