+++ src/client/src/assets/styles/primevueOverrides.scss
@@ -1,3 +1,5 @@
+@use '@/assets/styles/variables.scss';
+
 div.p-panel {
   border-style: none;
 }
