+++ src/client/src/components/views/dashboard/performers/PerformersGrid.vue
@@ -72,7 +72,7 @@ import api from '@/api/api'
 import { tryHandleUnsuccessfulResponse } from '@/api/apiResponseHandlers'
 import type { PerformerRequest } from '@/api/performers/performerRequest'
 import type { PerformerResponse } from '@/api/performers/performerResponse'
-import GridActions from '@/components/data/GridActions.vue'
+import GridActions from '@/components/data/grid-actions/GridActions.vue'
 import ListItem from '@/components/data/ListItem.vue'
 import DeleteDialog from '@/components/dialogs/DeleteDialog.vue'
 import FormDialog from '@/components/dialogs/FormDialog.vue'
