+++ src/client/src/components/views/dashboard/schedules/SchedulesGrid.vue
@@ -87,7 +87,7 @@ import { tryHandleUnsuccessfulResponse } from '@/api/apiResponseHandlers'
 import type { AudienceResponse } from '@/api/audiences/audienceResponse'
 import type { PerformerResponse } from '@/api/performers/performerResponse'
 import type { ScheduleResponse } from '@/api/schedules/scheduleResponse'
-import GridActions from '@/components/data/GridActions.vue'
+import GridActions from '@/components/data/grid-actions/GridActions.vue'
 import DeleteDialog from '@/components/dialogs/DeleteDialog.vue'
 import FormDialog from '@/components/dialogs/FormDialog.vue'
 import ScheduleForm from '@/components/form/requestForms/ScheduleForm.vue'
