+++ src/client/src/components/views/dashboard/schedules/TimeslotsGrid.vue
@@ -93,7 +93,7 @@ import type { ScheduleResponse } from '@/api/schedules/scheduleResponse'
 import { PerformanceType } from '@/api/schedules/timeslots/performanceType'
 import type { TimeslotRequest } from '@/api/schedules/timeslots/timeslotRequest'
 import type { TimeslotResponse } from '@/api/schedules/timeslots/timeslotResponse'
-import GridActions from '@/components/data/GridActions.vue'
+import GridActions from '@/components/data/grid-actions/GridActions.vue'
 import ListItem from '@/components/data/ListItem.vue'
 import DeleteDialog from '@/components/dialogs/DeleteDialog.vue'
 import FormDialog from '@/components/dialogs/FormDialog.vue'
