+++ src/client/src/utils/guards.ts
@@ -1,0 +1,3 @@
+export const throwIfUndefined = (obj?: any) => {
+  if (obj == undefined) throw new Error('Object was undefined where it must be defined.')
+}
